Fix: Statement ID to return the complete ID instead of just guid (#915)
## Description
In a previous PR we introduced the change #904 to only return the guid
when the `DatabricksResultSet.getStatementId` is called . This change
breaks the executeAsync flow, because when we try to execute an async
query from one statement and then poll for status using the other
statement we use `StatementId` as the identifier for the query.

Currently the type of client is determined by the format of the
StatementId. `guid|secret` statement id means it is a thrift client and
`guid` statement id means it is a SEA client.

So when we execute a query using the first connection in thrift mode and
then poll for the status using the second connection we get an error
because of the incorrect statement id format, the second connection
assumes the client to get SEA whereas the actual connection url
describes the client as thrift and this breaks the application

## Testing
- Manually tested
- Added Integration tests for executeAsync 

`NO_CHANGELOG=true`