Fix JDBC compliance of DatabricksStatement#setMaxRows(0) (#901)
## Description
According to the JDBC interface contract of `Statement#setMaxRows`, a
value of 0 must be interpreted as "no limit". Currently passing a value
of 0 literally means 0 rows will be returned. Instead, the
`DEFAULT_RESULT_ROW_LIMIT` seems to capture the semantics of "no limit",
if I'm reading the code correctly, so I used that in the PR.

## Testing
Tests have been added to check that setting max rows of 0 results in the
`DEFAULT_RESULT_ROW_LIMIT` to be set.

## Additional Notes to the Reviewer
<!-- Share any additional context or insights that may help the reviewer
understand the changes better. This could include challenges faced,
limitations, or compromises made during the development process.
Also, mention any areas of the code that you would like the reviewer to
focus on specifically. -->

---------

Co-authored-by: Roman Borschel <roman@cluvio.com>
Co-authored-by: Vikrant Puppala <vikrantpuppala@gmail.com>
Co-authored-by: Vikrant Puppala <vikrant.puppala@databricks.com>