Add Thrift statement-Id in debug logs, needed as we don't have enough logs in error cases (#934)
## Description
<!-- Provide a brief summary of the changes made and the issue they aim
to address.-->

Added Thrift statement-Id in debug logs for error scenarios and also as
soon as they are available. The idea is to better debug issues by
identifying the statement-Id in case of errors.

## Testing
<!-- Describe how the changes have been tested-->


## Additional Notes to the Reviewer
<!-- Share any additional context or insights that may help the reviewer
understand the changes better. This could include challenges faced,
limitations, or compromises made during the development process.
Also, mention any areas of the code that you would like the reviewer to
focus on specifically. -->

NO_CHANGELOG=true