[PECOBLR-764] compute new telemetry details if absent from export errorlogs (#935)
## Description
In current code during an exception scenario the exportFailureLogs
expects that the TelemetryCollector would already have statement details
for the given statementId. But that is not true hence making this
consistent with other method behaviour where we initialise a new
instance of statementDetails.

## Testing
Attaching test file with example details
1. Without connection close command
2. With connection close command


[closed_connection_command_0.txt](https://github.com/user-attachments/files/21729464/closed_connection_command_0.txt)

[closed_connection_command_1.txt](https://github.com/user-attachments/files/21729465/closed_connection_command_1.txt)


## Additional Notes to the [Reviewer]
NO_CHANGELOG=true


[closed_connection_command_1.txt](https://github.com/user-attachments/files/21729463/closed_connection_command_1.txt)

---------

Signed-off-by: Nikhil Suri <nikhil.suri@databricks.com>