Fix null statementID in telemetry logs + make thread daemon (#938)
## Description
- Currently telemetry wont be sent for cases where operationHandle is
not present, this PR fixes that. Example instance where our telemetry
would be useless : https://databricks.atlassian.net/browse/PECOBLR-359
- Also made telemetry client async : this avoids JVM to be alive even
after main program completes

## Testing
<!-- Describe how the changes have been tested-->

## Additional Notes to the Reviewer
The daemon introduces a known bug that will be fixed in a later PR
https://databricks.atlassian.net/browse/PECO-2716
- We have set the telemetry thread as daemon thread in JDBC. But, if the
connection is closed and program is ended, the thread may not be able to
send telemetry logs. There are 2 solutions discussed within the team on
this (but we would need benchmarking/ analyse the pros cons on this to
conclude) :
    - Make close telemetry client flush sync in nature
    - shutdown hook added to the executor
- A new HTTP Client may be created on the conn.close because HTTP Client
corresponding to the connectionContext is removed but the telemetry task
may not be complete by then(This will be resolved by fixing the above
issue 1.a. But we will have to finalize this too)

NO_CHANGELOG=true