Fix passing of encoded userAgent string in BITool+ Add system configuration in every telemetry request (#940)
## Description
- The intention of appName in telemetry is to understand BI tools using
our driver. And we do not want that to break our UserAgent. This causes
DBeaver to hang as it has a app name as `DBeaver%2F25.1.4.202508031529`
(and encoded strings are not allowed in our userAgent check). This PR
fixes that.
- To create dashboards, we need system configuration in every telemetry
row. This PR also adds that.

## Testing
- Added unit tests
- Also tested manually using DBeaver Client

## Additional Notes to the Reviewer

- The userAgent is sent for all customers, and it may not be appropriate
for us to send info like processName and appName as part of this.
Moreover, UserAgent is supposed to be fixed for a JVM. The `unknown`
string issue on thrift table will be addressed by telemetry now. (i.e.,
we will do a join between two tables to identify unknowns)

- BI tool Error that is fixed in this PR : 

<img width="812" height="21" alt="Screenshot 2025-08-14 at 4 42 33 PM"
src="https://github.com/user-attachments/assets/77a3c01d-0306-4535-95da-67ae73ecc82e"
/>