Empty schema to null (#942)
## Description
<!-- Provide a brief summary of the changes made and the issue they aim
to address.-->
Origin:
com.databricks.sql.hive.thriftserver.DatabricksJdbcDialectTestSuite in
runtime/Spark fails because the JDBC URL looks like
`jdbc:databricks://sample-host.18.azuredatabricks.net:9999/;`. Note the
empty string between the last slash and semi-colon. When parsing JDBC
URL, we look for a schema entry between last slash (if present) and
semi-colon. If it is empty, we should make it to null.

## Testing
<!-- Describe how the changes have been tested-->
- Unit test
- com.databricks.sql.hive.thriftserver.DatabricksJdbcDialectTestSuite

## Additional Notes to the Reviewer
<!-- Share any additional context or insights that may help the reviewer
understand the changes better. This could include challenges faced,
limitations, or compromises made during the development process.
Also, mention any areas of the code that you would like the reviewer to
focus on specifically. -->