Added a  config file to toggle release details (#968)
## Description

https://drive.google.com/a/databricks.com/open?id=1wVnv_pHMbCsaLRWPtBl__OVjeCFSj5N0qJx9Ekp_UEU
To have continous development in team, we are introducing a release
config in repo. git scripts will read the config to identify if a
release is in progress or not. This check will enforce blocking a PR
merge automatically instead relying in manual communication.

## Testing
This will be tested once we write git script to use this config. For now
this is unused code

## Additional Notes to the Reviewer
Release master will have to first raise PR with updated config before
and after release process is completed.

Signed-off-by: Nikhil Suri <nikhil.suri@databricks.com>