Added optional EnableSQLValidationForIsValid property for isValid() to allow users to enable server-side SQL validation (#961)
## Description
<!-- Provide a brief summary of the changes made and the issue they aim
to address.-->
Added SQL execution based isValid check gated by a connection property
EnableSQLValidationForIsValid. This allows users to enable server-side
SQL validation of isValid(). The validation is based on the execution of
[SELECT
VERSION()](https://databricks.atlassian.net/browse/ES-1361851?focusedCommentId=6176366)
query for optimised latency.
This is an optional property. The default behaviour is not changed for
isValid()
## Testing
<!-- Describe how the changes have been tested-->
Tested manually using a large query to check if the timeout functions
Added unit test
well.
## RELATED TICKETS AND DOCUMENTS
[PECOBLR-772](https://databricks.atlassian.net/browse/PECOBLR-772)
[ES-1361851](https://databricks.atlassian.net/browse/ES-1361851)

[PECOBLR-772]:
https://databricks.atlassian.net/browse/PECOBLR-772?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[ES-1361851]:
https://databricks.atlassian.net/browse/ES-1361851?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ