[PECOBLR-849] Fix JDBC Comparator workflow  (#986)
## Description
<!-- Provide a brief summary of the changes made and the issue they aim
to address.-->
The comparator workflow was not working since it was not able to resolve
the merge conflicts. Updated the workflow to apply the incoming changes
always in case of merge conflicts.
## Testing
<!-- Describe how the changes have been tested-->
Tested the command locally
## Additional Notes to the Reviewer
<!-- Share any additional context or insights that may help the reviewer
understand the changes better. This could include challenges faced,
limitations, or compromises made during the development process.
Also, mention any areas of the code that you would like the reviewer to
focus on specifically. -->
NO_CHANGELOG=true