fix: struct metadata parsing for nested decimal fields (#982)
## Description
When parsing Struct<Decimal(10,20)>, the driver currently raises an
exception:
```
  java.lang.ArrayIndexOutOfBoundsException: Index 1 out of bounds for length 1
      at com.databricks.jdbc.api.impl.MetadataParser.parseStructMetadata(MetadataParser.java:30)
      at com.databricks.jdbc.api.impl.ComplexDataTypeParser.parseToStruct(ComplexDataTypeParser.java:104)
      at com.databricks.jdbc.api.impl.ComplexDataTypeParser.convertValueNode(ComplexDataTypeParser.java:131)
      at com.databricks.jdbc.api.impl.ComplexDataTypeParser.parseToArray(ComplexDataTypeParser.java:73)
```
This occurs because the current implementation does not expect nested
comma's in datatypes.
By tracking the Depth of the `()` characters, we can successfully parse
complex datatypes.

Issue:
https://github.com/databricks/databricks-jdbc/issues/981

## Testing
Unit tests expanded and local testing done.

## Additional Notes to the Reviewer
This is a bug in the implementation of the MetaDataParser and is pretty
direct to re-produce and correct.

---------

Signed-off-by: josecsotomorales <josecsmorales@gmail.com>
Co-authored-by: Jayant Singh <jayant.singh@databricks.com>