Fixed ResultSetMetadata for SEA inline mode (#1010)
## Description
<!-- Provide a brief summary of the changes made and the issue they aim
to address.-->
For SEA this type is parsed later using a function. So we need not strip
the base type here. So reverted the code to use stripTypeName instead.
But the stripTypeName function isn't correct either since it will
incorrectly convert ARRAY<DECIMAL(10,2)> to ARRAY<DECIMAL and later
parse it incorrectly.
## Testing
<!-- Describe how the changes have been tested-->
Tested the SQL smoke test locally
## Additional Notes to the Reviewer
<!-- Share any additional context or insights that may help the reviewer
understand the changes better. This could include challenges faced,
limitations, or compromises made during the development process.
Also, mention any areas of the code that you would like the reviewer to
focus on specifically. -->
NO_CHANGELOG=true