Fixed stripTypeName function (#1012)
## Description
<!-- Provide a brief summary of the changes made and the issue they aim
to address.-->
Updated the stripTypeName function to strip the type only when it
contains the close bracket in the end. This would parse
ARRAY<DECIMAL(10,2)> correctly.
## Testing
<!-- Describe how the changes have been tested-->
Added unit test
## Additional Notes to the Reviewer
<!-- Share any additional context or insights that may help the reviewer
understand the changes better. This could include challenges faced,
limitations, or compromises made during the development process.
Also, mention any areas of the code that you would like the reviewer to
focus on specifically. -->

[PECOBLR-951](https://databricks.atlassian.net/browse/PECOBLR-951)

[PECOBLR-951]:
https://databricks.atlassian.net/browse/PECOBLR-951?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
NO_CHANGELOG=true