Added support for enableMultipleCatalogSupport param (#1025)
## Description
<!-- Provide a brief summary of the changes made and the issue they aim
to address.-->
Added support for the connection param of enableMultipleCatalogSupport.
The default value is kept to 1.
The behaviour remains the same when the param is enabled. When the param
is disabled, the behaviour is as follows:
- getCatalogs() returns only the current catalog
- Metadata operations are done on the current catalog
- setCatalog() method does nothing (doesn't execute USE CATALOG)
- The SQL query of USE CATALOG can still be executed.

To maintain the current catalog, we make the query to get the current
catalog everytime when this param is disabled. 
 
## Testing
<!-- Describe how the changes have been tested-->

## Additional Notes to the Reviewer
<!-- Share any additional context or insights that may help the reviewer
understand the changes better. This could include challenges faced,
limitations, or compromises made during the development process.
Also, mention any areas of the code that you would like the reviewer to
focus on specifically. -->
[PECOBLR-938](https://databricks.atlassian.net/browse/PECOBLR-938)


[PECOBLR-938]:
https://databricks.atlassian.net/browse/PECOBLR-938?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ