Add external scopes to U2M workflow (#1035)
## Description
- We had issues with SCOPES which we fixed in SDK
[[Link](https://github.com/databricks/databricks-sdk-java/commit/11237a2ed9775c1e09b5c8e0978f52f403213415)],
it works for M2M use-cases. But, the current ask is to now resolve this
for U2M too.


## Testing
- Added unit tests

## Additional Notes to the Reviewer
<!-- Share any additional context or insights that may help the reviewer
understand the changes better. This could include challenges faced,
limitations, or compromises made during the development process.
Also, mention any areas of the code that you would like the reviewer to
focus on specifically. -->

NO_CHANGELOG=true